requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
ollama==0.1.7
gradientai==1.4.0
streamlit==1.28.0
google-generativeai
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.2
